last clean-up before merge, was a fun branch, this feels like goodbye
