+++ Main/Main.java
@@ -2,7 +2,6 @@ package Main;
 
 import JDBC.ConnectionProvider;
 import javafx.application.Application;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
@@ -12,8 +11,8 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("new__sample.fxml"));
-        loader.setController(new NewController(new ConnectionProvider()));
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("sample.fxml"));
+        loader.setController(new Controller(new ConnectionProvider()));
         Parent root = loader.load();
         primaryStage.setTitle("MyTabb");
         primaryStage.setScene(new Scene(root, 900, 500));
