--- Main/new__sample.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane fx:id="anker" prefHeight="706.0" prefWidth="1088.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Label layoutX="165.0" layoutY="14.0" text="Who tf are you?!">
-         <font>
-            <Font size="37.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
