+++ Main/sample.fxml
@@ -1,35 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="588.0" prefWidth="615.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" >
+<AnchorPane fx:id="anker" prefHeight="706.0" prefWidth="1088.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
    <children>
-      <Label layoutX="158.0" layoutY="48.0" prefHeight="81.0" prefWidth="300.0" text="Who are you?">
+      <Label layoutX="165.0" layoutY="14.0" text="Who tf are you?!">
          <font>
-            <Font size="43.0" />
+            <Font size="37.0" />
          </font>
       </Label>
-      <Label layoutX="81.0" layoutY="166.0" text="Voornaam" />
-      <Label layoutX="261.0" layoutY="166.0" text="Achternaam" />
-      <TextField fx:id="voornaam" layoutX="81.0" layoutY="182.0" />
-      <TextField fx:id="achternaam" layoutX="261.0" layoutY="182.0" prefHeight="26.0" prefWidth="209.0" />
-      <Button fx:id="launch" layoutX="490.0" layoutY="182.0" mnemonicParsing="false"  text="Let's go" />
-      <ImageView fitHeight="253.0" fitWidth="384.0" layoutX="339.0" layoutY="248.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../properties/images/chiro.jpg" />
-         </image>
-      </ImageView>
-      <ImageView fitHeight="253.0" fitWidth="271.0" layoutX="22.0" layoutY="248.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../properties/images/Jules.jpg" />
-         </image>
-      </ImageView>
-      <Label layoutX="139.0" layoutY="542.0" text="brougth to you by the great mind of Jules Vervaeke " />
    </children>
 </AnchorPane>
