+++ errorHandling/MyError.java
@@ -1,7 +1,5 @@
 package errorHandling;
 
-import JDBC.ConnectionProvider;
-import Main.Controller;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
